# doc-files/kypo-training-swagger-open-api.yaml
...
swagger: "2.0"
info:
  version: "Version: 2.3.6"
  title: "KYPO Training - API Reference"
host: "localhost:8083"
basePath: "/kypo-rest-training/api/v1"
...
schemes:
- "http"
- "https"
paths:
...
  /training-instances/{instanceId}/training-runs:
    get:
      tags:
      - "Training instances"
      summary: "Get all training runs of specific training instance"
      description: "This can only be done by organizer of training instance or administrator"
      operationId: "findAllTrainingRunsByTrainingInstanceId"
      produces:
      - "application/json"
      parameters:
      - name: "instanceId"
        in: "path"
        description: "Training Instance Id"
        required: true
        type: "integer"
        format: "int64"
      - name: "isActive"
        in: "query"
        description: "If only active or not active training runs should be returned."
        required: false
        type: "boolean"
      - in: "body"
        name: "body"
        description: "Pagination support."
        required: false
        schema:
          $ref: "#/definitions/Pageable"
      - name: "fields"
        in: "query"
        description: "Fields which should be returned in REST API response"
        required: false
        type: "string"
      responses:
        200:
          description: "The training runs have been found."
          schema:
            $ref: "#/definitions/TrainingRunRestResource"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
...
  /visualizations/commands/training-instances/{instanceId}:
    get:
      tags:
      - "Command Visualizations"
      summary: "Get all commands executed in the given training instance associated\
        \ by training runs"
      description: ""
      operationId: "getAllCommandsByTrainingInstance"
      produces:
      - "application/json"
      parameters:
      - name: "instanceId"
        in: "path"
        description: "Training Instance ID"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "All commands for the given training instance found"
          schema:
            type: "object"
        404:
          description: "All commands for the given training instance not found"
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /visualizations/commands/training-instances/{instanceId}/aggregated:
    get:
      tags:
      - "Command Visualizations"
      summary: "Get aggregated correct/incorrect commands executed during the given\
        \ training runs."
      description: ""
      operationId: "getAggregatedCommandsForOrganizer"
      produces:
      - "application/json"
      parameters:
      - name: "instanceId"
        in: "path"
        description: "Training Instance ID"
        required: true
        type: "integer"
        format: "int64"
      - name: "runIds"
        in: "query"
        description: "Training Run IDs"
        required: true
        type: "List of numbers"
        items:
          type: "integer"
          format: "int64"
      - name: "correct"
        in: "query"
        description: "If correct or incorrect commands should be returned."
        required: true
        type: "boolean"
      - name: "mistakeTypes"
        in: "query"
        description: "Mistake types"
        required: false
        type: "List of strings"
        items:
          type: "string"
          enum:
          - "SYNTAX_UNKNOWN_COMMAND"
          - "SYNTAX_INVALID_ARGUMENTS"
          - "SYNTAX_INVALID_OPTION"
          - "SYNTAX_MISSING_OPTION_PARAM"
          - "SYNTAX_INVALID_OPTION_PARAM"
          - "SEMANTIC_UNKNOWN_IP"
          - "SEMANTIC_WRONG_IP"
      responses:
        200:
          description: "Aggregated correct/incorrect commands found."
          schema:
            $ref: "#/definitions/VisualizationInfoDTO"
        404:
          description: "Aggregated correct/incorrect commands for the given training\
            \ runs not found."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /visualizations/commands/training-runs/{runId}:
    get:
      tags:
      - "Command Visualizations"
      summary: "Get commands executed during the given training run."
      description: ""
      operationId: "getAllCommandsByTrainingRun"
      produces:
      - "application/json"
      parameters:
      - name: "runId"
        in: "path"
        description: "Training Run ID"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Commands for the given training run found."
          schema:
            $ref: "#/definitions/VisualizationInfoDTO"
        404:
          description: "Commmands for the given training run not found."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /visualizations/commands/training-runs/{runId}/aggregated:
    get:
      tags:
      - "Command Visualizations"
      summary: "Get aggregated correct/incorrect commands executed during the given\
        \ training runs."
      description: ""
      operationId: "getAggregatedCommandsForTrainee"
      produces:
      - "application/json"
      parameters:
      - name: "runId"
        in: "path"
        description: "Training Run ID"
        required: true
        type: "integer"
        format: "int64"
      - name: "correct"
        in: "query"
        description: "If correct or incorrect commands should be returned."
        required: true
        type: "boolean"
      - name: "mistakeTypes"
        in: "query"
        description: "Mistake types"
        required: false
        type: "List of strings"
        items:
          type: "string"
          enum:
          - "SYNTAX_UNKNOWN_COMMAND"
          - "SYNTAX_INVALID_ARGUMENTS"
          - "SYNTAX_INVALID_OPTION"
          - "SYNTAX_MISSING_OPTION_PARAM"
          - "SYNTAX_INVALID_OPTION_PARAM"
          - "SEMANTIC_UNKNOWN_IP"
          - "SEMANTIC_WRONG_IP"
      responses:
        200:
          description: "Aggregated correct/incorrect commands found."
          schema:
            $ref: "#/definitions/VisualizationInfoDTO"
        404:
          description: "Aggregated correct/incorrect commands for the given training\
            \ runs not found."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /visualizations/commands/training-runs/{runId}/all:
    get:
      tags:
      - "Command Visualizations"
      summary: "Get all commands executed in the given Training Run"
      description: ""
      operationId: "getAllCommandsByRun"
      produces:
      - "application/json"
      parameters:
      - name: "runId"
        in: "path"
        description: "Training Run ID"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "All commands for the given Training Run found"
          schema:
            type: "object"
        404:
          description: "All commands for the given Training Run not found"
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /visualizations/graphs/reference/training-definitions/{definitionId}:
    get:
      tags:
      - "Command Visualizations"
      summary: "Get reference graph for the given training definition."
      description: ""
      operationId: "getReferenceGraphForDesigner"
      produces:
      - "application/json"
      parameters:
      - name: "definitionId"
        in: "path"
        description: "Training definition ID."
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Reference graph found."
          schema:
            $ref: "#/definitions/VisualizationInfoDTO"
        404:
          description: "Reference graph for the given training definition not found."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /visualizations/graphs/reference/training-instances/{instanceId}:
    get:
      tags:
      - "Command Visualizations"
      summary: "Get reference graph for the given training instance."
      description: ""
      operationId: "getReferenceGraphForInstructor"
      produces:
      - "application/json"
      parameters:
      - name: "instanceId"
        in: "path"
        description: "Training instance ID."
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Reference graph found."
          schema:
            $ref: "#/definitions/VisualizationInfoDTO"
        404:
          description: "Reference graph for the given training instance not found."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /visualizations/graphs/reference/training-runs/{runId}:
    get:
      tags:
      - "Command Visualizations"
      summary: "Get reference graph for the given training run."
      description: ""
      operationId: "getReferenceGraphForTrainee"
      produces:
      - "application/json"
      parameters:
      - name: "runId"
        in: "path"
        description: "Training run ID"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Reference graph found."
          schema:
            $ref: "#/definitions/VisualizationInfoDTO"
        404:
          description: "Reference graph for the given training run not found."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /visualizations/graphs/summary/training-instances/{instanceId}:
    get:
      tags:
      - "Command Visualizations"
      summary: "Get summary graph for the given training instance."
      description: ""
      operationId: "getSummaryGraph"
      produces:
      - "application/json"
      parameters:
      - name: "instanceId"
        in: "path"
        description: "Training definition ID"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Summary graph found."
          schema:
            $ref: "#/definitions/VisualizationInfoDTO"
        404:
          description: "Summary graph for the given training instance not found."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /visualizations/graphs/trainee/training-runs/{runId}:
    get:
      tags:
      - "Command Visualizations"
      summary: "Get reference graph for the given training run."
      description: ""
      operationId: "getReferenceGraphForTrainee"
      produces:
      - "application/json"
      parameters:
      - name: "runId"
        in: "path"
        description: "Training definition ID"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Reference graph found."
          schema:
            $ref: "#/definitions/VisualizationInfoDTO"
        404:
          description: "Reference graph for the given training run not found."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
...
  /visualizations/training-instances/{instanceId}/training-runs:
    get:
      tags:
      - "Command Visualizations"
      summary: "Get finished training runs of the training instance."
      description: ""
      operationId: "findFinishedTrainingRunsByTrainingInstanceId"
      produces:
      - "application/json"
      parameters:
      - name: "instanceId"
        in: "path"
        description: "Training Instance ID"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Finished runs in the given training instance found."
          schema:
            $ref: "#/definitions/VisualizationInfoDTO"
        404:
          description: "Finished runs in the given training instance not found."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
...
securityDefinitions:
  bearerAuth:
    description: "Obtain bearer token from OIDC provider and enter the token in the\
      \ following field. Use format 'Bearer {token}'"
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  ApiError:
    type: "object"
    properties:
      timestamp:
        type: "integer"
        format: "int64"
        example: "1574062900 (different for each type of exception)"
        description: "The time when the exception occurred"
      status:
        type: "string"
        example: "404 Not found (different for each type of exception)."
        description: "The HTTP response status code"
        enum:
        - "CONTINUE"
        - "SWITCHING_PROTOCOLS"
        - "PROCESSING"
        - "CHECKPOINT"
        - "OK"
        - "CREATED"
        - "ACCEPTED"
        - "NON_AUTHORITATIVE_INFORMATION"
        - "NO_CONTENT"
        - "RESET_CONTENT"
        - "PARTIAL_CONTENT"
        - "MULTI_STATUS"
        - "ALREADY_REPORTED"
        - "IM_USED"
        - "MULTIPLE_CHOICES"
        - "MOVED_PERMANENTLY"
        - "FOUND"
        - "MOVED_TEMPORARILY"
        - "SEE_OTHER"
        - "NOT_MODIFIED"
        - "USE_PROXY"
        - "TEMPORARY_REDIRECT"
        - "PERMANENT_REDIRECT"
        - "BAD_REQUEST"
        - "UNAUTHORIZED"
        - "PAYMENT_REQUIRED"
        - "FORBIDDEN"
        - "NOT_FOUND"
        - "METHOD_NOT_ALLOWED"
        - "NOT_ACCEPTABLE"
        - "PROXY_AUTHENTICATION_REQUIRED"
        - "REQUEST_TIMEOUT"
        - "CONFLICT"
        - "GONE"
        - "LENGTH_REQUIRED"
        - "PRECONDITION_FAILED"
        - "PAYLOAD_TOO_LARGE"
        - "REQUEST_ENTITY_TOO_LARGE"
        - "URI_TOO_LONG"
        - "REQUEST_URI_TOO_LONG"
        - "UNSUPPORTED_MEDIA_TYPE"
        - "REQUESTED_RANGE_NOT_SATISFIABLE"
        - "EXPECTATION_FAILED"
        - "I_AM_A_TEAPOT"
        - "INSUFFICIENT_SPACE_ON_RESOURCE"
        - "METHOD_FAILURE"
        - "DESTINATION_LOCKED"
        - "UNPROCESSABLE_ENTITY"
        - "LOCKED"
        - "FAILED_DEPENDENCY"
        - "UPGRADE_REQUIRED"
        - "PRECONDITION_REQUIRED"
        - "TOO_MANY_REQUESTS"
        - "REQUEST_HEADER_FIELDS_TOO_LARGE"
        - "UNAVAILABLE_FOR_LEGAL_REASONS"
        - "INTERNAL_SERVER_ERROR"
        - "NOT_IMPLEMENTED"
        - "BAD_GATEWAY"
        - "SERVICE_UNAVAILABLE"
        - "GATEWAY_TIMEOUT"
        - "HTTP_VERSION_NOT_SUPPORTED"
        - "VARIANT_ALSO_NEGOTIATES"
        - "INSUFFICIENT_STORAGE"
        - "LOOP_DETECTED"
        - "BANDWIDTH_LIMIT_EXCEEDED"
        - "NOT_EXTENDED"
        - "NETWORK_AUTHENTICATION_REQUIRED"
      message:
        type: "string"
        example: "The IDMGroup could not be found in database (different for each\
          \ type of exception)."
        description: "The specific description of the ApiError."
      errors:
        type: "array"
        example: "[The requested resource was not found (different for each type of\
          \ exception).]"
        description: "The list of main reasons of the ApiError."
        items:
          type: "string"
      path:
        type: "string"
        example: "/kypo-rest-user-and-group/api/v1/groups/1000 (different for each\
          \ type of exception)."
        description: "The requested URI path which caused error."
    description: "Superclass for classes ApiEntityError and ApiMicroserviceError"
...
  Pageable:
    type: "object"
    properties:
      offset:
        type: "integer"
        format: "int64"
      sort:
        $ref: "#/definitions/Sort"
      paged:
        type: "boolean"
      unpaged:
        type: "boolean"
      page_number:
        type: "integer"
        format: "int32"
      page_size:
        type: "integer"
        format: "int32"
...
  Sort:
    type: "object"
    properties:
      empty:
        type: "boolean"
      sorted:
        type: "boolean"
      unsorted:
        type: "boolean"
...
  TrainingRunDTO:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        example: 1
        description: "Main identifier of training run."
      start_time:
        type: "string"
        format: "date-time"
        example: "2016-10-19 10:23:54+02"
        description: "Date when training run started."
      end_time:
        type: "string"
        format: "date-time"
        example: "2022-10-19 10:23:54+02"
        description: "Date when training run ends."
      event_log_reference:
        type: "string"
      state:
        type: "string"
        example: "ALLOCATED"
        description: "Current state of training run."
        enum:
        - "RUNNING"
        - "FINISHED"
        - "ARCHIVED"
      sandbox_instance_ref_id:
        type: "string"
        description: "Reference to the received sandbox."
      participant_ref:
        description: "Reference to participant of training run."
        $ref: "#/definitions/UserRefDTO"
      event_logging_state:
        type: "boolean"
        example: true
        description: "Boolean to check whether event logging works."
      command_logging_state:
        type: "boolean"
        example: true
        description: "Boolean to check whether command logging works."
      has_detection_event:
        type: "boolean"
        example: true
        description: "Boolean to check whether the run has any detection events logged"
    description: "The act, or a recording, of performing actions during training from\
      \ a perspective of one concrete participant."
  TrainingRunRestResource:
    type: "object"
    properties:
      content:
        type: "array"
        description: "Retrieved Training Runs from databases."
        items:
          $ref: "#/definitions/TrainingRunDTO"
      pagination:
        description: "Pagination including: page number, number of elements in page,\
          \ size, total elements and total pages."
        $ref: "#/definitions/Pagination"
    description: "Content (Retrieved data) and meta information about REST API result\
      \ page. Including page number, number of elements in page, size of elements,\
      \ total number of elements and total number of pages"
...
  VisualizationInfoDTO:
    type: "object"
    properties:
      training_definition_id:
        type: "integer"
        format: "int64"
        example: 1
        description: "Main identifier of training definition."
      training_definition_title:
        type: "string"
        example: "TrainingDefinition2"
        description: "A name of the training/game (e.g., Photo Hunter) ."
      training_definition_estimated_duration:
        type: "integer"
        format: "int64"
        example: 5
        description: "Estimated time it takes to finish runs created from this definition."
      levels:
        type: "array"
        description: "Information about all levels in training definition."
        items:
          $ref: "#/definitions/AbstractLevelVisualizationDTO"
    description: "A model includes necessary information about training definition\
      \ which is needed for visualizations."